Troubleshooting v2: moneda en Hardware, biblioteca de Implementacion editable, aprobacion para Implementacion, validaciones enteras y rediseno de documentos
- Selector de moneda (COP/USD) al crear items de Hardware, con totales agrupados por moneda
- Nueva pagina Biblioteca de Implementacion (CRUD de tecnologias/actividades) + PUT/DELETE en catalogo-implementacion
- Se saca Bolsa de Horas y Site Survey del abanico de Implementacion; Site Survey se agrega como tipo de servicio
- Aprobacion opcional (checkbox) para Implementacion, replicando el flujo de 6 estados que ya tenia Servicios
- Validacion de enteros positivos en cantidades (hardware, ingenieros/dias/plantas/factor equipos)
- Rediseno de excelCotizacion.js, excelBom.js y wordPropuestaTecnica.js siguiendo los formatos reales de Veolia y la oferta de referencia
- Campo de comentarios/notas en el BOM, incluido en el Excel consolidado

+++ backend/src/modules/boms/hardwareItems.routes.js
@@ -10,9 +10,9 @@ hardwareItemsRouter.use(authMiddleware);
 
 const agregarSchema = z.object({
   hardwareId: z.number().int().positive(),
-  cantidad: z.number().positive().default(1),
+  cantidad: z.number().int().positive().default(1),
 });
-const actualizarSchema = z.object({ cantidad: z.number().positive() });
+const actualizarSchema = z.object({ cantidad: z.number().int().positive() });
 
 async function assertBomAccess(req) {
   await bomsService.getBom(req.params.bomId, req.user);
